requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
db-dtypes==1.4.4
decorator==5.2.1
dlt==1.20.0
frozenlist==1.8.0
fsspec==2025.12.0
gcsfs==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
giturlparse==0.14.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-cloud-bigquery==3.38.0
google-cloud-bigquery-storage==2.35.0
google-cloud-core==2.5.0
google-cloud-storage==3.7.0
google-cloud-storage-control==1.8.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
httplib2==0.31.0
humanize==4.14.0
idna==3.11
jsonpath-ng==1.7.0
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.0
numpy==2.3.5
oauthlib==3.3.1
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pathvalidate==3.3.1
pendulum==3.1.0
pluggy==1.6.0
ply==3.11
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.2
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
requirements-parser==0.13.0
rich==14.2.0
rich-argparse==1.7.2
rsa==4.9.1
semver==3.0.4
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.45
sqlglot==28.0.0
tenacity==9.1.2
tomlkit==0.13.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.1
wheel==0.45.1
win-precise-time==1.4.2
yarl==1.22.0
